Attack Replay Generator
Using the C Programming Language, libdnet, 
and libpcap the goal was to replay an attack on a machine for which we have a tcpdump formatted binary file of 
all network traffic and a configuration file. This attack replay generator uses libdnet to inject all of the packets 
with the exception of the victim packets on the network and uses libpcap to read packets from the network or 
tcpdump log file. When you inject packets in the network the victim will send packets in response to your 
attempt to connect and my program reads the contents of the victim packets and modifies the packet(ack 
number, sequence number, size, etc) so that it can keep responding to the victim and establish/maintain a 
connection. Can replay three way handshakes, telnet sessions, finger sessions, and possibly more if updated to 
handle buffer overflows, ftp, and http sessions.