PotTYpicker
Using Java, JavaFX, VioletUML, and excel we created a desktop application called PotTY Picker. The application PotTY Picker is designed to give ratings and 
status updates on the restrooms at UTSA. It allowed for one to five star ratings, user review which were kept in 
an excel file, user accounts whose passwords and usernames were stored in an excel file, a map of the school 
with restroom locations, and an email function to request cleanings or restocking of restrooms.